requirements.txt
torch~=2.1.0
pyyaml~=6.0.1
tqdm~=4.66.1
pandas~=2.1.1
